Extract out clever link into concern
